fix(audit): F-12/F-34 widen nx*ny*nz to long long before multiply
FluidLBM and VOFSolver constructors compute num_cells_ via int*int*int.
For domains larger than ~536M cells (sqrt(INT_MAX) ≈ 46340 per dim^2)
the multiply silently wraps to negative or near-zero, breaking every
allocation that follows.

Phase-5 / Phase-1 are 8.25M cells — far from the limit. But future AMR
or multi-GPU subdomains could plausibly hit 1B+ total cells per node.

Cast first arg to long long; the int*long_long*long_long product is
safe; final assignment to int num_cells_ still narrows but at least
the underlying arithmetic is correct (and would be flagged by sanitizer
if ever wrapped negative).

Long-term TODO: change num_cells_ to long long across the codebase.

+++ src/physics/fluid/fluid_lbm.cu
@@ -55,7 +55,10 @@ FluidLBM::FluidLBM(int nx, int ny, int nz,
                    BoundaryType boundary_z,
                    float dt, float dx)
     : nx_(nx), ny_(ny), nz_(nz),
-      num_cells_(nx * ny * nz),
+      // F-12 (code-audit pass 1): widen to long long before multiply to avoid
+      // int overflow at >536M cells (e.g., AMR/multi-GPU domains). num_cells_
+      // is still int — assignment narrows but the multiply itself is safe.
+      num_cells_(static_cast<long long>(nx) * ny * nz),
       dt_(dt), dx_(dx),
       nu_physical_(kinematic_viscosity),
       rho0_(density),
